Device identity initialization improvements.
These updates take the first step forward in providing the
ability to configure a device with its identity.  If a device
identity is available on the root boot filesystem the identity
manager uses the id-set utility to load the identity into
NVram.

The identity manager extends the PCR identity register with the
measured boot status, the software state and the implementation
of the device identity.  The implementation is then reduced to a
hash which will be made available by the identity manager.

This process invalidates the ability of the platform to read its
NVram identity.