Implement and integrate IDmgr object.
The IPC object which was previously added to the source tree was
encapsulated in the IDmgr object to implement an object which
requests transaction services from the identity manager daemon.

The current service which is supported is the ability to return
the hash of the identity implementation and the identity key.
These are the components needed to implement the generation of
OTEDKS keys.

The identity manager daemon was updated to the IDmgr API to issue
transaction services.  Cleanups were also added to the signal
handling infrastructure to trap the SIGINT instruction and
cleanly exit.

The utils/Makefile directory was modified to change the library
link order to have HurdLib at the end of the list.  This was done
to address an issue with dependencies introduced by the IPC
object.