Implement IDtoken exchange in identity manager.
Support was added for the ->set_idtoken and ->get_idtoken methods
in the IDmgr object.  This provides support for obtaining a full
identity implement from the the manager.

The previous implementation of the object was short sighted since
it overlooked the fact that POSSUM needed the full implementation
of an identity in order to generate a POSSUM authentication
token.