Use identity manager to access identitys in possum.
The possum host and client modes were modified to use the IDmgr
object to obtain the host identities.  Functionality was
confirmed by removing the host identity from the configuration
image.

The security supervisor was modified to wait for proper startup
of the identity manager daemon.  This prevents failure in
userspace which starts behind the identity manager due to the
identity manager not being accessible.