Add quote generation and verification.
The ->verify and ->generate_quote methods were added to the
TPMcmd object. These methods provide the ability to verify a
machine status quote and to generate a reference quote for a
platform.

The tpm-cmd utility was updated to implement verify and
generate-quote functionality to exercise these methods.

Various cleanups were implemented across the TPMcmd method.