Updates to boot utilities.
The storage root key is stored in the /var/lib/tpm/system.data
file which requires that this file be managed as a configuration
element and thus loaded from the system configuration state.

The switch_root() function was updated to copy this file from
the /etc/conf directory to its final location.

In a software based tpm the software state is in
/var/lib/swtpm/00.permall.  The sboot tool also copies this file
from /etc/conf to its final location.

The boot utility was also modified to switch its real root
identity to the no privileged identity when it loads its
configuration files.