Add support for generation of PCR composite masks.
The ->pcrmask method was added to implement the creation of a
mask across a specified test of PCR registers.  This method is
variadic in order to specify a complete set of registers in a
single call.

Each call will clear the composite mask so the current mask will
be based on the last call to the function.