Use PCR mask status for quoting.
The ->quote and ->generate_quote methods were updated to use the
current PCR mask settings of their enclosing objects.

The tpm-cmd utility was modified to use the ->pcrmask method to
request that registers 10 and 17 be used for quote generation.

The verify function in the tpm-cmd utility was not properly
setting the return status which was corrected.