Add support for creation of an identity verified to IDtoken.
The ->to_verifier method was added to the IDtoken module in order
to support the reduction of an identity token into an identity
verifier.  The utils/gen-id-verifier was doing that with inline
code and due to coarseness of the ->reset method requred more
code then what an internal implementation does.

The gen-id-verifier utility was modified to carry out the
identity conversion using the new method.