Add support for ASN1 encoding of identify verifier.
There were numerous updates to the identity verifier object in
this update.  The major implementation change was to provide for
the ability to encode/decode a verifier in ASN1 format.  This
will be the preferred format for storing files on participating
devices.

The gen-id-verifier utility was also updated, significantly, to
create an identify verifier.