Place correct identity in verifier.
The ->set_identity method was placing the wrong element into the
id Buffer object in Ivy.  Only the reduced implementation of the
identity was being added.

The method requests conversion of the identity token into ASN1
form and loads the entire encoded identity into the id element of
the verifier.