Add support for IVY client identification.
The possum driver utility was updated to use the recently
introduced IVY support to load configuration information for its
counter-parties.

This sets the stage for validating machine quotes which will be
incorporated in upcoming updates.

The PossumPacket method needs to be upgraded to separate the
replay nonce from the hardware quote nonce.  Currently they are
loaded into a single buffer which makes the possum host and
client mode implementations a bit clumsy.