Add utilities for generating TXT policies.
This commit introduces two utilities used for generating the
launch control and trusted boot policies in an Intel Trusted
Execution Technology enviornment.

The gen-lcp-policy generates and loads into NVram the launch
control policy which measures the tboot hypervisor and its
command-line.

the gen-tb-policy generates and loads into NVram the trusted boot
policy which measures the kernel command-line and its parameters.