Implement proper extension of identity.
The initial implementation of the Identity object failed to
follow correct extension strategy in the ->create method.  This
update corrects the extension method.

The precursor to the extension of the organizational identity
with the identity of the entity is to first extend the anonymizer
for the identity/class with the hash of the identifier for the
entity.

This hash is then used to extend the organizational identity to
form the final identity rooted in the organizational identity
heirarchy.