Add error handling to identity engine and client.
The ->set_error method was added to the identity engine object in
order to implement an accessor method for returning an error to
the client which is requesting generation of an identity.  The
identity generation engine was updated to use this to return an
error to the client.

The identity generation client interprets an error response and
responds accordingly.