Add interactive support to gen-id utility.
This commit adds support for specifying that the gen-id utility
run in 'interactive' mode by specifying the -I command-line
switch to the utility.

In this mode the utility continually requests identity decriptor
information and uses that information to query the identity
generation engine for the identity derived from this component
information.

Properly supporting this functionality entailed the addition of
an additional identity type to the IDengine_identity enumerated
class.