Add support for network based identity generation.
The major purposes of these changes is to provide the general
framework for a network based client/server architecture for creating
identities.

The idsvr utility runs and communicates via shared memory IPC
with an identity generation.  It accepts remote connection
attempts and receives an ASN1 encoded version of the IPC
structure and submits that to the local identity generation.
The output IPC structure is ASN1 encoded and returned to the
client

The fetch-id utility is the client utility for accessing this
infrastructure and encodes the identity generation request for
transmission to the network server.