Update identity client/server.
The dependencies for the fetch-id and idsvr utilities were
updated in the Makefile in order to force rebuilds of the
binaries if the objects they are dependent on change.

Both the client and server were updated to use port 10903 as the
service port.

The identity server was updated to better handle termination in
the identity generation child process.