Verify that identity implementation needs to be reduced.
An error was noted when the identity token was being checked out
from the identity manager.  This was secondary to the fact that
the identity manager stores the device identity in reduced form.

The ->set_identity method was modified to check the size of the
identity implementation.  Reduction of the identity
implementation is only executed if the implementation is in full
form.