Add support for extracting public key.
The ->get_pubkey method was added to the TPMcmd object which
implements the ability to extract the public portion of a key
which is stored in the persistent system keystore.  The key is
returned in DER encoded form suitable for being used to create an
identity verifier.

The tpm-cmd utility was updated to add the get-pubkey command
which exercises this functionality.