Make identity NVram be dependent on PCR register 15.
This update modifies the locking of the identity into NVram.
The id-set utility was modified to specify that a read of the
identity NVram index should only be allowed at the current value
of register PCR 15.

In order to tie the read of the device identity to the hardware
and software state register 15 is extended with registers 10, 17
and 18 prior to the extension with the device identity being
done.

The extension of the identity register previously occurred with a
concantentation of the current values of the hardware and
software registers.  The extendsions are now done sequentially.

This induced a timing race since additional TPM transactions are
required.  The number of wait iterations was extended to a total
of 60 seconds for the identity manager to come online.