Implement selectable reverse DNS lookups.
Running in an imbedded environment without DNS causes each
connection to experience a five second delay while the reverse
lookup on the connecting IP address timesout.  This commit adds a
boolean variable to the Duct object state information which
selects whether or not reverse lookups should be done.

By default reverse lookups are disabled.  The ->do_reverse method
is provided to modify this setting.