Properly handle multiple POSSUM partner identities.
Setting up a POSSUM based IPsec tunnel was failing when multiple
identity partners were specified.  This was being caused by the
find_client() function failing to properly initialize itself on
each identity evalation cycle.

The ->reset method was added to the Ivy object in order to reset
the effective identity state of the Ivy object to that of the
identity partner being evaluated.

The possum utility was updated to take advantage of this
functionality and clean up various other object state
initialization in the find_client() loop.