Fix problems with multi-tenant IPsec tunnels.
Implementing multiple IPsec tunnels was failing secondary to two
issues which this commit addresses.

The find_clients() function was reading the identity verification
files for each participating peer using root privileges.  This
was causing the .ivy files to be added to the measurement state
of the system thus perturbing the measurement of the system.

Including the measurement of a peer system in the system
measurement results in a cyclic dependency which cannot be
resolved.  This issue was addressed by temporarily changing the
real uid of possum to the non-privileged user while the .ivy
files were read.

This commit also removes reading of the 'client_list' variable
from the possum configuration file.  This is no longer needed
since reading the .ivy files replaces the functionality of
the client list file.