Implement support for forwarding identity reciprocation requests.
This commit implements multiple changes to support handling
requests to reciprocate an identity.  The recip-svr receives a
request over an IPsec tunnel from a client with an identity
to be reciprocated.  That identity is passed to the local
identity generation and the result is returned to the client.

This update also implements support for returning the status of
the local identity generation request.  There appears to be
issues with that support returning a false positive failure so
actual use is disabled at this time.