Add standalone net configuration tool.
It was becoming apparent that configuring network interfaces from
within the security supervisor (init) was inflexible,
particularly in the presence of systems with multiple interfaces.
The net-config utility was implement to provide basic
functionality for defining network interface addresses and
routes.

The net-config utility takes a mode arguement, either -I or -R,
which selects configuriong an interface or a route.  Three
arguements follow which specify the parameters for the interface
and route.