Implement support for generating root and config seal files.
This update provides infrastructure which allows the system boot
loader to generate a sealed password file if a password file is
found in /boot/root.pwd on the root filesystem.  This release
does not support the use of PCR register 10 secondary to issues
with the boot command not yet being deterministic with respect to
the execution graph leading up to the point where the load-image
utility is called.