Updates to identity generation server/client.
The changes embodied in this update implement a functional
identity generation server and graphical client.  This reflects
the version which was delivered to Noridian for testing.

It appears as if it is not a straight forward exercise to imbed a
Java keystore in a .jar file.  The code was modified to set the
keystore property from within the graphical client.  This means
the client can be started with the following command:

	java -cp idclient.jar idclient

As long as the idclient.jks file is in the current working
directory.