Fully initialize sigaction structures.
The sigaction structure was not being properly initialized prior
to a call to set signal handlers.  While this did not manifest in
our codebase experience in an alternate codebase indicated this
could result in random segmentation faults.

This commit explicitly sets the flags variable in the sigaction
structure and uses the sigemptyset call to initialize the signal
handler mask.