Update boot and identity set utilities.
This update implements a small set of fixes needed to allow the
system to be boot-strapped in a vTPM environment.

The delay after the trousers daemon was started was increased
since the daemon appears to be starting slower then in either a
hardware or software simulator environment.  This delay needs to
be replaced with some type of test to determine if the daemon is
operational as the timing method does not appear to be
deterministic.

The id-set utility was modified to not generate an error if the
NVram region is not defined.  Instead the assumption is made that
the reason for the error was that the region did not exist.  This
needs to be changed to an explicit test for the presence of the
region.