Makefile updates for idtest utility.
The idtest target was not properly specifying all of the object
files needed to successfully link the binary.  Update the idtest
target to include the necessary object files.

The idtest binary was not being included in the TOOLS variable.
Add the name of the binary to the variable so that the clean:
target removes the binary.