Add identity name arguement to IDmgr lookups.
In order to support subsequent updates which will allow alternate
identities to be selected from the manager the accessor methods
were modified to add a service name arguement in the form of a
String object.

By default all of the existing infrastructure which uses the
IDmgr object was modified to select the device identity with the
'device' identity name.
In order to support subsequent updates which will allow alternate identities
to be