Implement support for management of multiple identities.
The changes implemented in this commit provide the infrastructure
needed to support the management of multiple identities.

A fixed size array of structures is used to implement the pool of
identities under management.  This is currently of a fixed size
of 10 identities which is sufficient for the current
implementation.  This does not need to be potentially implemented
as a dynamically sizable resource pool.

Each structure element in the array contains a String object and
an IDtoken object.  The String object names the identity which is
maintained in reduced form suitable for key generation.