Test and correct request of multiple identities.
The idtest utility was modified to accept a -i command line
arguement to specify an alternate identity name.  This was
implemented in order to exercise the functionality which was
added in the last commit.

This functionality resulted in finding and correcting an issue
with requesting identities.  The first issue was the fact the
String object which was loaded with the name was not being reset.
This resulted in a steadily progressive string length as multiple
requests were processed.

The comparison function which was determining whether or not the
names matched was changed to use a strcmp function in order to
correct an issue with the partial match of the two device names.