Implement common build framework.
This commit implements system to provide a generic build
configuration.  This is in form of a file named Makefile.mak
which provides common build directives which are currently being
repliaced in multiple suborinate Makefiles.

This is only a preliminary step in what needs to be extensive
work on the build configuration.