Remove LIBS variable from LDFLAGS variable.
The LDFLAGS variable is positioned at the front of the link
commands which appears to cause issues with the libraries not
being included into the symbol resolution change.  As a result
links failed due to unresolved references.

The LIBS variable now needs to be added to each of the link
commands in the subordinate makefiles.