Add frame for EDI transaction processor.
This update provides the framework for a application component
which implements the secure exchange of Electronic Data
interchange (EDI) messages.  EDI transactions or messages are
used in multiple industries for information exchange.

EDI transactions consist of one or more lines of each text, each
of which is terminated with a tilde (~).  Individual fields
within a line are seperated with asterisks (*).  The first field
in the line denotes the component in the message which the line
represents.

The EDI transaction exchanger consists, at this time, of three
separate daemons:

	edi-transmitter
	edi-engine
	edi-receiver

The edi-transmitter daemon accepts ASCII EDI transactions as
input.

The transmitter encodes the transaction in ASN1 format and passes
the transaction structure to the edi-engine which is designed to
run on the HUI node of a Chicken Cruncher.  The edi-engine
contacts the local identity server to request the information
needed to generate an OTEDKS key which is used to encrypt the EDI
payload.

The encrypted payload is returned to the edi-transmitter which
sends the payload over an encrypted PossumpPipe or IPsec connection
to the edi-receiver.

The edi-receiver hands the encrypted transaction structure to the
edi-engine which is running on the HUI node of the Chicken
Cruncher which is running the receiver.  The payload is decrypted
and the resulting structure is passed back to the edi-receiver
for final delivery to the destination devices accessible to the
receiver.