Generic fixes and updates to identity manager object and daemon.
The previous update which added the ability to lookup the
identity hash and key illustrated a number of deficiencies in the
identit management infrastructure.  This commit addresses those
deficiencies and thus has a number of disparate changes and
fixes.

The identity manager was not properly processing a request
generated by the ->get_id_key method.  The manager was
unconditionally expecting the use of the ->get_idtoken method.

This was addressed by adding the ->get_idtype method which
returns the enumerated type which specifies the type of request
made by the client.

A regression was found wherein the ->get_id_key method was
unconditionally erasing the entire ipc data structure.  Since
that structure contained the process id of the daemon this
resulted in subsequent requests after the first to fail.  A new
internal function named _clear_ipc was called to encapsulate the
clearing of all the data elements in the IPC structure.

The locking code was improved to prevent stale locks from being
left which could cause the identity manager to deadlock.