Add object for managing EDI transactions.
This update is a staging commit which provides support for the
EDIpacket object which encapsulates the information needed to
transport an EDI transation.  The object in gross context
contains the encryption time for the object, the type of payload
encased in the object and the payload itself.

The object is designed to take as input and generate as output an
ASN1 encoded object containing this information.  This object
implements the packet which is sent across the network.