Integrate EDIpacket support into transaction processing system.
This commit integrates the EDIpacket object introduced in the last
commit into the transaction processing chain.  This change
results in all transactions now being ASN1 encoded.

The EDIpacket infrastructure is also leveraged to encrypt the
payload object before it is transported across the network.

As of this commit all of the infrastructure is in place for
end-to-end encrypted transmission of EDI transactions.  This
framework will be used to build forward additional functionality.