Rationalize capture of transaction times.
The previous implementation for timing the end-to-end time needed
to complete an EDI transaction did not cope properly with the
time rolling past a micro-second boundary.  This resulted in
artifically large transaction times when a request was executed
near this boundary.

This commit captures the start and end times as the current epoch
time in micro-seconds stored in a 64 bit integer.  The
transaction time is expressed, in micro-seconds, as the
difference between these two values.