Implement support for ephemeral EDI transaction keying.
This is a somewhat large and involved update whose primary
purpose is to implement support for insuring that EDI
transactions have Perfect Forward Security (PFS) protection when
they are transmitted over the wire between the EDI transmitter
and receiver.

These changes were also necessitated by the fact that OTEDKS key
scheduling has an inherent one second time granularity.  This
means that if EDI transactions are transmitted more rapidly then
one per second there would be multiple transactions which were
encrypted with the same key and initialization vector.

The most immediate implication of this is that observation of
multiple encrypted transactions would allow the potential
discernment of transactions which were identical and which were
transmitted during the same OTEDKS keying period.  While this
would be a highly unlikely scenario in practice it was apparent
with the test harness which uses transmission of the same EDI
transaction.

All of these issues are addressed through the implementation of a
PFS constraint.  Upon startup the EDI transmitter opens a
connection with the EDI receiver and uses an Elliptic Curve
Diffie Hellman key exchange to implement a shared secret between
the transmitter and receiver.

The EDI transmitter and receiver send the shared secret to the
EDI service encryption engine and then destroy the secret so the
only entities with long term posession of the key are the
engines.

The shared secret is hashed and used as the key to an HMAC which
takes as its input the OTEDKS generated key and initialization
vector.  The hash of the shared secret is iterated for each
EDI transaction received which results in a differnet
initialization vector and key regardless of the OTEDKS epoch key
and IV.

The two-node security model is preserved since posession of the
ephemeral key does not convey any knowledge with respect to
the initialization vector and key used since the privacy of both
entities are protected by the first pre-image resistance of the
HMAC.