Add ability to send multiple EDI transactions per conection.
The previous implementation of send-edi had the option to implement
a EDI transaction 'batch' given a common input file.  This update
implements the ability to send either the single or batched input
multiple times in the context of a given connection to the EDI
transmission system.

The number of transmission repetitions is given with the '-r'
command-line arguement.