Add initial framework for PossumPipe object.
This update provides the initial development framework for a
PossumPipe object.  This object will provide a client/server
implementation for a communications conduit which is protected by
mutual remote identity attestation between a client and server.

The object will compile but is not functional at this commit
point.

A test harness named PossumPacket_test is also committed but is
not included in the build process at this time.