Fix shared key generation.
Code review indicated an issue with how the shared key was being
generated based on the Elliptic Curve Diffie Hellman key which
was exchanged.  This commit fixes the regression in both the
PossumPipe object and the possum utility.  The possum utility
will be migrated to use PossumPipe when development of the latter
is completed.

The generation of the shared secret is based on the following:

	Shared = HMACdh(client_nonce ^ server_nonce)

Where HMACdh is a SHA256 based HMAC using the ECDH generated
secret as the keying value.

The error int he implementation was that the following formula
was being used:

	Shared = HMACdh(dh)

Where the input to HMAC was the shared key rather then the XOR of
the nonces provided by the participants in the connection.