Update shared key generation.
The primary focus of this update is move the shared keys which
are generated into the object state structure.  This is needed to
allow packet sends and receive access to keying information.

The generation of the shared keys was also modified to generate
two separate keys.  The first key will be used to provide
initialization vectors and encryption authentication keys with
the second providing the encryption key.

The following formulas are used for key generation:

	Shared1 = sha256(HMAC_dhkey(client_nonce ^ host_nonce))

	Shared2 = sha256(HMAC_dhkey(host_nonce || client_nonce))

Where dhkey is the shared secret obtained from the ECDH key
exchange and the HMAC is SHA256 based.

The PossumPipe_test utility was modified to add an information
message on successful server startup.