Implement packet send/receive and ASN1 encoding.
The update provides the following methods to the object:

	->send_packet
	->receive_packet

These methods are designed to be the standard interfaces for
initiating the transmission and reception of data from a remote
endpoint.  The type of the packet to be sent is specified as an
arguement to the ->send_packet method.

Internally ASN1 encoding was implemented to encode the type of
the packet and its actual payload.

The test harness, PossumPipe_test, was updated to exercise this
new functionality.