Integrate software integrity measurement into authentication checksum.
This commit implements support for binding the packet authentication
checksum to the software integrity state of the platform.  The
HMAC key used to generate the checksum is computed from the
following:

	Key = HMACimakey(shared1high)

Where imakey is the SHA256 extension hash of the IMA file
template hash which is generated by the SoftwareStatus object.

The shared1high input to the HMAC is the top 128 bits of the
current extension of the shared1 key negotiated at connection
startup.

The ->packet_send method computes the current local IMA extension
hash on generation of each patcket authentication checksum.  The
remote software status which is attested at connection start is
stored in the connection object state and is used to generate the
authentication key for packet reception.

In this model a change in the integrity status of the communication
counter-party will cause a receive packet checksum failure.  The
decision to initiate a failure at this level is to prevent the
ability of a remote attacker to generate packets which could be
used to probe the remote decryption state.

The decision to compute the software integrity from the /sys
based IMA measurement pseudo-files was used rather then a TPM
reference quote due to the expense of obtaining reference quotes.
The two parties in the communications contract will have verified
the software measurement status by virtue of the platform
reference quotes which were exchanged at platform startup.

A connection re-keying event will re-validate the platform
measurement and will thus provide ongoing verification that the
IMA status remains constant.