Add ->extend method.
The changes in this update implement the ability to 'extend' the
current hash value with an arbitrary amount of input.  This
extension operation is the analogue of a TPM PCR register
extension.

The data to be used in the extension is supplied in a Buffer
object.  The SHA256 hash of this material is computed and concantenated
to the existing hash value.  A new hash value is computed over
these two combined values and becomes the new hash value of the
object.