Implement transmit/receive data extension registers.
This commit implements two SHA256 registers which represent a
continuously updated extension of the data which has been transmitted
by the local client and received from the remote communication
counter-party.  This strategy was influenced by the TPM security model
where Platform Configuration Registers (PCR's) are used to provide a
summary of the modification history of security sensitive material.

The extension values in these registers were then used to derive the
packet encryption key.  The formula for generating the key is as
follows:

	Kencryption_transmit = HMACecdh(SHA256_sent)

	Kencryption_receive = HMACecdh(SHA256_received)

Where HMACecdh is a SHA256 based HMAC using as a key the continually
hashed version of the initial shared secret from connection keying.
The input to the HMAC is either the extension hash of all data
transmitted up to that date or all data which was received to date,
dependent on whether a packet is being transmitted or received.

The root values for the registers are based on a hash of the two
shared secrets which are generated at connection setup time.  The
order of the hashing is inverted so the transmit and receive extension
registers have different initial values.

This model means that each transmitted packet should receive a unique
and distinct encryption key.  An attacker would have to have knowledge
of the entire transmitted data stream in order to re-create the
encryption keys in addition to the root shared secret.

In addition this prevents a third party from injecting a packet into
the communication stream since the reception of the packet would
poison the ability of the receiver to decrypt subsequent
communications.