Implement explicit EOF generation and detection.
This update provides the infrastructure needed for a client to
generate a write record size of 0 to indicate an explicit end of
transmission indication.

A method named ->eof was implemented to interrogate the status of
the flag which is set when a receiver detects that a
counter-party indicated an end-of-transmission condition.

All of this work was implemented to provide a more deterministic
method of indicating when a connection was being gracefully taken
down as opposed to an error condition being generated.