Re-enable dropping privileges when reading .ivy files.
Debugging the application with user level privileges had required
disabling the code which changes the real and effective userid to
non-root privileges.

Unfortunately that caused a rather long and tedious debugging
session to determine why connects were failing in a virtual
environment.  It will be better to debug the application with the
privileges it has in a production environment then run the risk
of having separate code for testing and production.