Add debug and error location code generation macros.
Debugging in a measured environment was becoming almost untenable
without strict error location tagging.  There is also the need to
begin the process of restricting output to debugging environments
only.

This update provides the following code generation macros:

	ERR
	SAY

The ERR macro takes a single arguement which is the action to be
executed after an output message is generated indicating the
location of where the fault occurs.

The SAY macro is conditionally compiled by the NAAAIM_DEBUG
define and takes a single arguement which is a message to be
printed.  The message is prevised with the function name in which
the macro was used.