Add per packet nonces.
This update implements the generation and inclusion of per packet
random payloads.  At connection initiation time a random block of
data equal to twice the blocksize of the encryption algorithm is
generated and hashed into a SHA256 object.

Four random bytes are also generated to implement a uint32_t
value to server as a seed for the srand() function to seed the
random number generation uniquely for each security context.

At packet send time a random number is generated, from which the
lower five bits are masked off to generate a random value in the
range of 0-31 bytes.  This random number of bytes is copied into
the nonce payload of the packet.

A second random number is generated from which the lower four
bits are masked off to to generate a number in the range of 0-15.
This number, plus one, is used as a count to re-hash the SHA256
object which serves as the nonce payload field.