Functionality update for RSAkey object.
This update provides preliminary support for allowing the RSAkey
object to be used with hardware key support.  The test utility
was also upgraded to provide a basic framework for testing the
hardware support as well as standard file based keys.

The hardware support was added throught the OpenSSL engine
support.  The ->init_engine method was added to support the
specification and configuration of a hardware engine to implement
the RSA key privimites.  This method is called with a pointer to
an array of strings which specify the commands needed to carry
out engine configuration.

While not necessarily needed the Makefile was supplemented with
an additional variable, SSLCRYPTO, which provides a definition
which links only the crypto library.  The previous method of
linking the library used a somewhat blunt approach of linking the
SSL library which is not necessarily a requirement for those
systems only using symbols from the encryption library.