Convert EDI receiver/transmitter to use PossumPipe conduit.
This commit is a staging point for future development and
refinement for the EDI receiver/transmitter to send transactions
through a security context implemented with a PossumPipe object.

It is important to note that at the current point the
edi-receiver fails on an initial connection when run in a Xen
virtual machine.  This does not seem to affect connections which
are run on 'bare-metal'.

The current hypothesis on this problem is that it may be due to
the process catching a signal while waiting on a read I/O
although that is only supposition at this point.

There is code which is commented out in edi-receiver.c which
allows the received to be run in a 'one-shot' mode which allows
the system to function properly in a virtual machine.