Send packet size and payload atomically in Duct object.
In its previous implementation the Duct object transmitted the
payload packet length followed by the actual payload.  The
implementation was naive in that the use of two separate write
commands resulted in two separate TCP exchanges.

The ->send_Buffer method was modified to use writev to coalesce
the two writes into a single TCP exchange.  Two separate IO
vectors were defined, the first contains the payload length with
the payload being defined in the second vector.

The impact of this change was significant.  In a complete
localhost implementation the total time for an EDI transaction
was reduced from 95,000 micro-seconds to 17,000 micro-seconds
which equates to a five fold decrease in transmission time.