Implement receive buffering.
The ->receive_Buffer method was using a somewhat primitive
implementation for slurping in the contents of the payload.  The
contents of the payload was being read one byte at a time and
then aggregated into the receiving Buffer object.

While the Linux system call pathway is very fast this is
decidedly sub-optimal for larger payload sizes.

A 1024 byte buffer is now allocated in the Duct state object
which is used by the ->receive_Buffer method to ingress larger
quantities of the payload in a single system call.  The payload
receive size is batched into multiple 1024 units with a final
read being conducted to pick up residual data which is not
contained in an integral block unit.