Initial commit of device hotplug manager.
In order to respond to the insertion of the USB hardware key
device the ability to create and remove USB device nodes is
needed.  This commit provides the initial framework for a daemon
named hotplug which manages this process.

This commit is only a staging commit and while it compiles the
utility is still not functional.