Complete implementation of hotplug daemon.
This update expands on the infrastructure which was introduced in
the last commit.  At this development point the hotplug daemon is
functional and properly handles addition and removal of a Yubikey
device.

The ability to run the daemon in debug mode was added via the -d
command-line switch.  This causes actional insertion and removal
events to be printed.