Add support for initializing connected devices.
This update provides support for creating device nodes for
devices which have been connected and initialized prior to
starting of the hotplug manager.

It should now be possible to start the pcscd card manager after
this utility has executed and have the daemon recognize a device
which was extant at boot.  After this the hotplug manager will
support the daemons hotplug detection system.